requirements.txt
# LangChain
langchain
langchain-core
langchain-openai
langchain-community

# LlamaIndex
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-readers-file
llama-index-vector-stores-chroma

# Vector Store
chromadb

# PDF
pypdf
pdfplumber

# Data
pandas
numpy

# Tokens
tiktoken

# HTTP
httpx
requests

# OpenAI
openai

# Utils
python-dotenv
python-dateutil

# UI
streamlit